Moved all feature specific ngrx stuff and interface stuff to video-games feature module
