+++ src/app/interfaces/loadable/index.ts
@@ -1,0 +1,1 @@
+export * from './loadable';
