+++ src/app/video-games/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './platforms';
+export * from './video-game-listing';
