+++ src/app/video-games/interfaces/platforms/index.ts
@@ -1,0 +1,1 @@
+export * from './platforms.interface';
