+++ src/app/video-games/interfaces/video-game-listing/index.ts
@@ -1,0 +1,2 @@
+export * from './video-game-filters.interface';
+export * from './video-game-listing.interface';
