+++ src/app/video-games/interfaces/video-game-listing/video-game-listing.interface.ts
@@ -1,4 +1,4 @@
-import {ILoadable} from '../loadable/loadable';
+import {ILoadable} from '../../../interfaces/loadable';
 import {
   IVideoGame, videoGameMatchesSearchQuery,
   videoGameMatchesFilters
