+++ src/app/video-games/store/platforms/platforms.effects.ts
@@ -2,9 +2,9 @@ import {Injectable} from '@angular/core';
 import {Effect, Actions} from '@ngrx/effects';
 import {Observable} from 'rxjs/Observable';
 
-import {createAction} from '../create-action';
-import {PlatformsService} from '../../video-games/services/platforms.service';
 import {PlatformsStore} from './platforms.store';
+import {createAction} from '../../../store/create-action';
+import {PlatformsService} from '../../services/platforms.service';
 
 @Injectable()
 export class PlatformsEffects {
