+++ src/app/video-games/store/reducers.ts
@@ -1,0 +1,2 @@
+export * from './platforms/platforms.reducer';
+export * from './video-game-listing/video-game-listing.reducer';
