+++ src/app/video-games/video-games.module.ts
@@ -1,6 +1,7 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {RouterModule} from '@angular/router';
+import {EffectsModule} from '@ngrx/effects';
 
 import {
   VideoGamesContainerComponent
@@ -9,13 +10,16 @@ import {
 import {PlatformsService, VideoGamesService} from './services';
 import {VideoGameListingModule} from './listing/listing.module';
 import {VideoGameDetailModule} from './detail/detail.module';
+import {PlatformsEffects, VideoGameListingEffects} from '../video-games/store/effects';
 
 @NgModule({
   imports: [
     CommonModule,
     RouterModule,
     VideoGameListingModule,
-    VideoGameDetailModule
+    VideoGameDetailModule,
+    EffectsModule.run(PlatformsEffects),
+    EffectsModule.run(VideoGameListingEffects),
   ],
   declarations: [
     VideoGamesContainerComponent
